smart_devops_recommender/backend/requirements.txt
requests
httpx
weaviate-client
polars
uvicorn
fastapi